+++ example/lib/main.dart
@@ -1,13 +1,21 @@
 import 'package:flutter/material.dart';
 import 'package:flame/game.dart';
-import 'game/platformer_game.dart';
+import 'game/simplified_game.dart';
+import 'game/json_game.dart';
 
 void main() {
   runApp(const MyApp());
 }
 
-class MyApp extends StatelessWidget {
+class MyApp extends StatefulWidget {
   const MyApp({super.key});
+
+  @override
+  State<MyApp> createState() => _MyAppState();
+}
+
+class _MyAppState extends State<MyApp> {
+  bool useJsonFormat = true;
 
   @override
   Widget build(BuildContext context) {
@@ -18,7 +26,21 @@ class MyApp extends StatelessWidget {
         primarySwatch: Colors.blue,
         useMaterial3: true,
       ),
-      home: GameWidget(game: PlatformerGame()),
+      home: Scaffold(
+        body: GameWidget(
+          key: ValueKey(useJsonFormat),
+          game: useJsonFormat ? JsonGame() : SimplifiedGame(),
+        ),
+        floatingActionButton: FloatingActionButton.extended(
+          onPressed: () {
+            setState(() {
+              useJsonFormat = !useJsonFormat;
+            });
+          },
+          label: Text(useJsonFormat ? 'JSON Format' : 'Simplified Format'),
+          icon: const Icon(Icons.swap_horiz),
+        ),
+      ),
     );
   }
 }