+++ lib/flame_ldtk.dart
@@ -8,10 +8,18 @@ library flame_ldtk;
 export 'src/models/ldtk_level.dart';
 export 'src/models/ldtk_entity.dart';
 export 'src/models/ldtk_intgrid.dart';
+export 'src/models/ldtk_json_models.dart';
 
 // Parsers
 export 'src/parsers/ldtk_super_simple_parser.dart';
+export 'src/parsers/ldtk_json_parser.dart';
+export 'src/parsers/ldtk_parser_utils.dart';
 
 // Components
 export 'src/components/ldtk_level_component.dart';
+export 'src/components/ldtk_json_level_component.dart';
 export 'src/components/ldtk_entity_component.dart';
+export 'src/components/ldtk_tile_layer_component.dart';
+
+// Additional models
+export 'src/models/ldtk_tile_layer.dart';